requirements.txt
bittensor>=9
setuptools>=70
requests>=2
numpy>=2
pandas==2.2.3
sqlalchemy==2
psycopg2-binary>=2.9.10
